now with double

+++ src/ch10/implicit/implicit.cpp
@@ -4,6 +4,6 @@ int main() {
 	char c{};
 	std::cout << "Enter char: ";
 	std::cin >> c;
-	std::cout << int(c) << '\n';
+	std::cout << double(c) << '\n';
 	return 0;
 }
